moved CaseInsensitiveModelBackend to a new backends.py file, updated settings to reflect the change
