+++ fantasy_football_app/backends.py
@@ -1,0 +1,24 @@
+from django.contrib.auth.backends import ModelBackend
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, get_user_model
+
+# -------------------------------------------------------------------------
+# ----- Authentication backend -----
+# -------------------------------------------------------------------------
+
+class CaseInsensitiveModelBackend(ModelBackend):
+    def authenticate(self, request, username=None, password=None, **kwargs):
+        UserModel = get_user_model()
+        try:
+            user = UserModel.objects.get(username__iexact=username)
+            if user.check_password(password):
+                return user
+        except UserModel.DoesNotExist:
+            return None
+
+    def get_user(self, user_id):
+        UserModel = get_user_model()
+        try:
+            return UserModel.objects.get(pk=user_id)
+        except UserModel.DoesNotExist:
+            return None
